2_StockSnaps_Backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
blinker==1.6.3
Brotli==1.1.0
celery==5.3.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.2.0
django-cors-headers==4.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.55.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.7
kombu==5.3.7
locust==2.17.0
MarkupSafe==2.1.3
matplotlib==3.7.2
msgpack==1.0.7
numpy==1.26.1
packaging==24.2
pandas==2.1.1
pillow==11.0.0
prompt_toolkit==3.0.48
psutil==5.9.6
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
redis==5.0.0
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
